рефакторинг

+++ include/mgf.hpp
@@ -21,10 +21,10 @@ namespace cry {
 
                 std::vector<uint8_t> mgfSeed(first, last);
 
-                mgfSeed.push_back((i & 0xFF000000) >> 24);
-                mgfSeed.push_back((i & 0x00FF0000) >> 16);
-                mgfSeed.push_back((i & 0x0000FF00) >> 8);
-                mgfSeed.push_back((i & 0x000000FF) >> 0);
+                mgfSeed.push_back(static_cast<uint8_t>((i & 0xFF000000) >> 24));
+                mgfSeed.push_back(static_cast<uint8_t>((i & 0x00FF0000) >> 16));
+                mgfSeed.push_back(static_cast<uint8_t>((i & 0x0000FF00) >> 8));
+                mgfSeed.push_back(static_cast<uint8_t>((i & 0x000000FF) >> 0));
 
                 hash(mgfSeed.begin(), mgfSeed.end(), result);
 
@@ -35,10 +35,10 @@ namespace cry {
             {
                 std::vector<uint8_t> mgfSeed(first, last);
 
-                mgfSeed.push_back((i & 0xFF000000) >> 24);
-                mgfSeed.push_back((i & 0x00FF0000) >> 16);
-                mgfSeed.push_back((i & 0x0000FF00) >> 8);
-                mgfSeed.push_back((i & 0x000000FF) >> 0);
+                mgfSeed.push_back(static_cast<uint8_t>((i & 0xFF000000) >> 24));
+                mgfSeed.push_back(static_cast<uint8_t>((i & 0x00FF0000) >> 16));
+                mgfSeed.push_back(static_cast<uint8_t>((i & 0x0000FF00) >> 8));
+                mgfSeed.push_back(static_cast<uint8_t>((i & 0x000000FF) >> 0));
 
                 std::vector<uint8_t> digest(hLen);
                 hash(mgfSeed.begin(), mgfSeed.end(), digest.begin());
